chore(theme): setup system providers
