+++ packages/c-theme-provider/src/theme.hooks.ts
@@ -1,0 +1,8 @@
+import { ThemeProviderProps } from '@chakra-ui/vue-next'
+import { computed, inject } from 'vue'
+
+/** Provides theme object in component context */
+export const useTheme = () => {
+  const theme = inject('$chakraTheme') as ThemeProviderProps
+  return theme
+}
